[DEV-2949] change expiry date usage
